+++ CHANGELOG.md
@@ -16,6 +16,10 @@ OpenNMT-tf follows [semantic versioning 2.0.0](https://semver.org/). The API cov
 
 ### Fixes and improvements
 
+## [1.10.1](https://github.com/OpenNMT/OpenNMT-tf/releases/tag/v1.10.1) (2018-10-15)
+
+### Fixes and improvements
+
 * Fix possible error when loading checkpoints without `--model_type` or `--model` after updating to a newer OpenNMT-tf version. The saved model description is now more future-proof regarding model class updates.
 * Fix embedding visualization when the vocabulary file is stored in the model directory or when a joint vocabulary is used
 * Improve encoder/decoder states compatibility check